Changing Leo's PR from a fork to a local PR.
This PR was created locally from the origina PR submitted by Leo Tan
from his fork of the project.  The build on Travis was failing to
authenticate to Docker Hub, so the PR is being recreated locally to
see if this addresses the authentication failure.

In addition, a minor change was made to the authentication line in
his original PR.

Leo has since been given collaborator rights to the original repo
so forking will not be necessary in the future.

On branch feature-add-drupal-based-php-with-apache

Changes to be committed:
	modified:   .travis.yml